Bump to k8s.io v0.30.3, controller-runtime v0.18.4
Bump the operator's vendored k8s.io/* packages to v0.30.3, and bump its
vendored sigs.k8s.io/controller-runtime package to v0.18.4, the latest
release, which uses the k8s.io/* v0.30.1 packages.

    cd operator
    go get k8s.io/api@v0.30.3 k8s.io/apiextensions-apiserver@v0.30.3 k8s.io/apimachinery@v0.30.3 k8s.io/client-go@v0.30.3 sigs.k8s.io/controller-runtime@v0.18.4
    go mod tidy
    go mod vendor

Update calls to controller.Watch to adapt to breaking changes in
controller-runtime.  Namely, the handler and predicate arguments have been
removed from Watch and are now specified as options on the source.
Additionally, the source.Kind type now has a type parameter, so it is
necessary to specify a client.Object type argument.

This commit is related to OCPBUGS-38102.

https://issues.redhat.com/browse/OCPBUGS-38102

* operator/controller/dnsnameresolver/controller.go (NewUnmanaged): Update
for the new controller.Watch and source.Kind type signatures.
* operator/go.mod: Bump.
* operator/go.sum:
* operator/vendor/*: Regenerate.